Update Float & Double.c

+++ C/Variable/Float & Double.c
@@ -2,7 +2,7 @@
 
 int main() {
 	double a = 10.5;
-	double b = 7;
+	int b = 7;
 	printf("a / b = %f", a / b);
 	return 0;
 }
